// src/main/webapp/src/app/rest/model/jiraIssueDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IssueExtraFieldsDTO } from './issueExtraFieldsDTO';
import { IssueHistoryDTO } from './issueHistoryDTO';
import { IssueFieldsDTO } from './issueFieldsDTO';


export interface JiraIssueDTO { 
    id?: string;
    key?: string;
    fields?: IssueFieldsDTO;
    histories?: Array<IssueHistoryDTO>;
    annotated?: IssueExtraFieldsDTO;
}